API: Director::handleRequest() is no longer static - use a Director service NEW: Add HTMLMiddlewareAware trait to HTTPApplication, Director, and RequestHandler NEW: Allow service specs to be passed to Director rules.
This refactor of the controller middlewares takes a service definition
approach rather than a static-method-and-config approach that Director
historically had.

The use of a trait for middleware means that the Middlewares array
property can be defined on RequestHandler, Director, and HTTPApplication
objects in the same way.
