+++ _config/requestprocessors.yml
@@ -1,16 +1,16 @@
 ---
 Name: requestprocessors
 ---
-SilverStripe\Control\Director:
-  middlewares:
-    TrustedProxyMiddleware: '%$SilverStripe\Control\TrustedProxyMiddleware'
-    AllowedHostsMiddleware: '%$SilverStripe\Control\AllowedHostsMiddleware'
-    SessionMiddleware: 'SilverStripe\Control\SessionMiddleware'
-    RequestProcessor: 'SilverStripe\Control\RequestProcessor'
-    FlushMiddleware: '%$SilverStripe\Control\FlushMiddleware'
-
-
 SilverStripe\Core\Injector\Injector:
+  SilverStripe\Control\Director:
+    properties:
+      Middlewares:
+        TrustedProxyMiddleware: '%$SilverStripe\Control\TrustedProxyMiddleware'
+        AllowedHostsMiddleware: '%$SilverStripe\Control\AllowedHostsMiddleware'
+        SessionMiddleware: '%$SilverStripe\Control\SessionMiddleware'
+        RequestProcessor: '%$SilverStripe\Control\RequestProcessor'
+        FlushMiddleware: '%$SilverStripe\Control\FlushMiddleware'
+
   SilverStripe\Control\AllowedHostsMiddleware:
     properties:
       AllowedHosts: "`SS_ALLOWED_HOSTS`"
